requirements.txt
discord.py
asyncio==3.4.3
pymysql
apscheduler
Pillow
